comment deletion security add-on

+++ routes/forum.js
@@ -261,7 +261,7 @@ router.post("/comment/delete", async (req, res) => {
     let user = await User.findOne({ access_token: req.query.access_token });
     try {
         if (user) {
-            await Comment.deleteOne({ _id: req.query.comment_id })
+            await Comment.deleteOne({ _id: req.query.comment_id, author: user._id });
             res.json({ success: true });
         } else {
             res.json({ success: false, error: "User not found." });
